+++ frontend/src/tests/Messenger.test.tsx
@@ -7,19 +7,9 @@ import {
   waitFor,
 } from "@testing-library/react";
 import Messenger from "../components/Messenger";
-import MessageBubble from "../components/MessageBubble";
-import { useEffect } from "react";
-import {
-  RecoilRoot,
-  RecoilState,
-  useRecoilState,
-  useRecoilValue,
-} from "recoil";
-import {
-  currentThreadNameState,
-  messageBubblesState,
-  sidebarState,
-} from "../components/atoms";
+import { RecoilRoot } from "recoil";
+import { messageBubblesState, sidebarState } from "../components/atoms";
+import { RecoilTestHelper, RecoilObserver } from "./RecoilStateHelper";
 jest.mock("../components/WebSocket-Context", () => ({
   useWebSocket: () => ({
     dispatchWebSocket: jest.fn(),
@@ -51,36 +41,6 @@ beforeEach(() => {
 });
 afterEach(cleanup);
 
-type RecoilObserverProps = {
-  node: RecoilState<any>;
-  onChange: Function;
-};
-
-const RecoilObserver = ({ node, onChange }: RecoilObserverProps) => {
-  const value = useRecoilValue(node);
-  useEffect(() => onChange(value), [onChange, value]);
-  return null;
-};
-
-type RecoilStateHelperProps = {
-  sidebarActivated: boolean;
-  messageBubbles: JSX.Element[];
-  currentThreadName: string;
-};
-const RecoilTestHelper = ({
-  sidebarActivated,
-  messageBubbles,
-  currentThreadName,
-}: RecoilStateHelperProps) => {
-  const [, toggleSideBar] = useRecoilState(sidebarState);
-  const [, addBubble] = useRecoilState(messageBubblesState);
-  const [, setCurrentThreadName] = useRecoilState(currentThreadNameState);
-  toggleSideBar(sidebarActivated);
-  addBubble(messageBubbles);
-  setCurrentThreadName(currentThreadName);
-  return <div></div>;
-};
-
 it("messenger test", async () => {
   render(
     <RecoilRoot>
@@ -103,18 +63,9 @@ it("messenger test", async () => {
   fireEvent.click(buttonSend);
 
   expect(addBubble).toBeCalled();
-  // expect(jsx_list.length).toBe(1);
 });
 
 it("empty input test", () => {
-  jsx_list = [
-    <MessageBubble
-      text="test"
-      sender="message-bubble-user"
-      prevSender={"message-bubble-volunteer"}
-    />,
-  ];
-
   render(
     <RecoilRoot>
       <RecoilTestHelper
@@ -132,14 +83,6 @@ it("empty input test", () => {
 });
 
 it("prevSender test", async () => {
-  jsx_list = [
-    <MessageBubble
-      text="test"
-      sender="message-bubble-user"
-      prevSender={"message-bubble-volunteer"}
-    />,
-  ];
-
   render(
     <RecoilRoot>
       <RecoilTestHelper
